Revert "test: Split raft uv tests that rely on mocking getaddrinfo"
This reverts commit 87e8e6bf62d44edbbf66394d203ba16daf4c0f7f.
