--- test/raft/integration/main_addrinfo.c
@@ -1,3 +1,0 @@
-#include "../lib/runner.h"
-
-RUNNER("addrinfo")
