+++ test/raft/integration/test_uv_tcp_connect.c
@@ -1,5 +1,6 @@
 #include "../../../src/raft.h"
 #include "../../../src/raft.h"
+#include "../lib/addrinfo.h"
 #include "../lib/heap.h"
 #include "../lib/loop.h"
 #include "../lib/runner.h"
@@ -125,6 +126,7 @@ static void *setUpDeps(const MunitParameter params[],
 {
     struct fixture *f = munit_malloc(sizeof *f);
     int rv;
+    SET_UP_ADDRINFO;
     SET_UP_HEAP;
     SETUP_LOOP;
     SETUP_TCP_SERVER;
@@ -142,6 +144,7 @@ static void tearDownDeps(void *data)
     TEAR_DOWN_TCP_SERVER;
     TEAR_DOWN_LOOP;
     TEAR_DOWN_HEAP;
+    TEAR_DOWN_ADDRINFO;
     free(f);
 }
 
@@ -185,6 +188,29 @@ TEST(tcp_connect, connectByName, setUp, tearDown, 0, NULL)
     char host_adress[256];
     sprintf(host_adress, "localhost:%d", TCP_SERVER_PORT);
     CONNECT(2, host_adress);
+    return MUNIT_OK;
+}
+
+/* Successfully connect to the peer by first IP  */
+TEST(tcp_connect, firstIP, setUp, tearDown, 0, NULL)
+{
+    struct fixture *f = data;
+    const struct AddrinfoResult results[] = {{"127.0.0.1", TCP_SERVER_PORT},
+                                             {"192.0.2.0", 6666}};
+    AddrinfoInjectSetResponse(0, 2, results);
+    CONNECT(2, "any-host");
+    return MUNIT_OK;
+}
+
+/* Successfully connect to the peer by second IP  */
+TEST(tcp_connect, secondIP, setUp, tearDown, 0, NULL)
+{
+    struct fixture *f = data;
+    const struct AddrinfoResult results[] = {{"127.0.0.1", .6666},
+                                             {"127.0.0.1", TCP_SERVER_PORT}};
+
+    AddrinfoInjectSetResponse(0, 2, results);
+    CONNECT(2, "any-host");
     return MUNIT_OK;
 }
 
@@ -295,6 +321,33 @@ TEST(tcp_connect, closeDuringConnectAbort, setUp, tearDownDeps, 0, NULL)
     munit_assert_int(rv, ==, 0);
     check.data = f;
     CONNECT_REQ(2, BOGUS_ADDRESS, 0, RAFT_NOCONNECTION);
+    /* Successfull DNS lookup will initiate async connect */
+    LOOP_RUN(1);
+    uv_check_start(&check, checkCb);
+    LOOP_RUN(1);
+    LOOP_RUN_UNTIL(&_result.done);
+    CLOSE_WAIT;
+    return MUNIT_OK;
+}
+
+/* The transport gets closed right after the first connection attempt failed,
+ * while doing a second connection attempt. */
+TEST(tcp_connect, closeDuringSecondConnect, setUp, tearDownDeps, 0, NULL)
+{
+    struct fixture *f = data;
+    struct uv_check_s check;
+    int rv;
+    const struct AddrinfoResult results[] = {{"127.0.0.1", .6666},
+                                             {"127.0.0.1", TCP_SERVER_PORT}};
+
+    AddrinfoInjectSetResponse(0, 2, results);
+
+    /* Use a check handle in order to close the transport in the same loop
+     * iteration where the second connection attempt occurs. */
+    rv = uv_check_init(&f->loop, &check);
+    munit_assert_int(rv, ==, 0);
+    check.data = f;
+    CONNECT_REQ(2, "any-host", 0, RAFT_CANCELED);
     /* Successfull DNS lookup will initiate async connect */
     LOOP_RUN(1);
     uv_check_start(&check, checkCb);
