+++ test/raft/integration/test_uv_tcp_listen.c
@@ -1,5 +1,6 @@
 #include "../../../src/raft.h"
 #include "../../../src/raft/byte.h"
+#include "../lib/addrinfo.h"
 #include "../lib/heap.h"
 #include "../lib/loop.h"
 #include "../lib/runner.h"
@@ -94,6 +95,7 @@ static void *setUpDeps(const MunitParameter params[],
                        MUNIT_UNUSED void *user_data)
 {
     struct fixture *f = munit_malloc(sizeof *f);
+    SET_UP_ADDRINFO;
     SET_UP_HEAP;
     SETUP_LOOP;
     SETUP_TCP;
@@ -106,6 +108,7 @@ static void tearDownDeps(void *data)
     TEAR_DOWN_TCP;
     TEAR_DOWN_LOOP;
     TEAR_DOWN_HEAP;
+    TEAR_DOWN_ADDRINFO;
     free(f);
 }
 
@@ -228,6 +231,83 @@ static MunitParameterEnum invalidTcpListenParams[] = {
 TEST(tcp_listen, invalidAddress, setUp, tearDown, 0, invalidTcpListenParams)
 {
     struct fixture *f = data;
+    LISTEN(RAFT_IOERR);
+    return MUNIT_OK;
+}
+
+/* Check success with addrinfo resolve to mutiple IP and first one is used to
+ * connect */
+TEST(tcp_listen, firstOfTwo, setUp, tearDown, 0, NULL)
+{
+    const struct AddrinfoResult results[] = {{"127.0.0.1", 9000},
+                                             {"127.0.0.2", 9000}};
+    struct fixture *f = data;
+    AddrinfoInjectSetResponse(0, 2, results);
+    LISTEN(0);
+    PEER_CONNECT;
+    PEER_HANDSHAKE;
+    ACCEPT;
+    return MUNIT_OK;
+}
+
+/* Check success with addrinfo resolve to mutiple IP and second one is used to
+ * connect */
+TEST(tcp_listen, secondOfTwo, setUp, tearDown, 0, NULL)
+{
+    const struct AddrinfoResult results[] = {{"127.0.0.2", 9000},
+                                             {"127.0.0.1", 9000}};
+    struct fixture *f = data;
+    AddrinfoInjectSetResponse(0, 2, results);
+
+    LISTEN(0);
+    PEER_CONNECT;
+    PEER_HANDSHAKE;
+    ACCEPT;
+    return MUNIT_OK;
+}
+
+/* Simulate port already in use error by addrinfo response contain the same IP
+ * twice */
+TEST(tcp_listen, alreadyBound, setUp, tearDown, 0, NULL)
+{
+    /* We need to use the same endpoint three times as a simple duplicate will
+     * be skipped due to a glib strange behavior
+     * https://bugzilla.redhat.com/show_bug.cgi?id=496300  */
+    const struct AddrinfoResult results[] = {
+        {"127.0.0.1", 9000}, {"127.0.0.1", 9000}, {"127.0.0.1", 9000}};
+    struct fixture *f = data;
+    AddrinfoInjectSetResponse(0, 3, results);
+    LISTEN(RAFT_IOERR);
+    return MUNIT_OK;
+}
+
+/* Error in bind first IP address */
+TEST(tcp_listen, cannotBindFirst, setUp, tearDown, 0, NULL)
+{
+    const struct AddrinfoResult results[] = {{"192.0.2.0", 9000},
+                                             {"127.0.0.1", 9000}};
+    struct fixture *f = data;
+    AddrinfoInjectSetResponse(0, 2, results);
+    LISTEN(RAFT_IOERR);
+    return MUNIT_OK;
+}
+
+/* Error in bind of second IP address */
+TEST(tcp_listen, cannotBindSecond, setUp, tearDown, 0, NULL)
+{
+    const struct AddrinfoResult results[] = {{"127.0.0.1", 9000},
+                                             {"192.0.2.0", 9000}};
+    struct fixture *f = data;
+    AddrinfoInjectSetResponse(0, 2, results);
+    LISTEN(RAFT_IOERR);
+    return MUNIT_OK;
+}
+
+/* Check error on general dns server failure */
+TEST(tcp_listen, resolveFailure, setUp, tearDown, 0, NULL)
+{
+    struct fixture *f = data;
+    AddrinfoInjectSetResponse(EAI_FAIL, 0, NULL);
     LISTEN(RAFT_IOERR);
     return MUNIT_OK;
 }
